float32 is faster
Former-commit-id: 7b02071379493e21af4bf9c1064062b6cfa3aed7
Former-commit-id: 42aacf4e00b78f8aa8973b63b6fec67dd025ef79

+++ JAxtar/annotate.py
@@ -1,6 +1,6 @@
 import jax.numpy as jnp
 
-KEY_DTYPE = jnp.float16
+KEY_DTYPE = jnp.float32
 ACTION_DTYPE = jnp.uint8
 HASH_POINT_DTYPE = jnp.uint32
 HASH_TABLE_IDX_DTYPE = jnp.uint8
